requirements.txt
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
h5py==3.8.0
ipython==8.12.2
jedi==0.18.2
joblib==1.2.0
matplotlib-inline==0.1.6
numpy==1.24.3
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wcwidth==0.2.6
